+++ test2/optimization.h
@@ -8,8 +8,8 @@
 #include "Utility.h"
 #include <limits>
 
-Matrix WMRA_Opt(int i, float JLA, float JLO, Matrix Jo, float detJo, Matrix dq, vector<double> delta, float dt, vector<double> cur);
+Matrix WMRA_Opt(int i, double JLA, double JLO, Matrix Jo, double detJo, Matrix dq, vector<double> delta, double dt, vector<double> cur);
 void WMRA_Jlimit(Matrix& qmin, Matrix& qmax);
-Matrix WMRA_Opt(Matrix Jo, float detJo, vector<double> dx, vector<double> q);
+Matrix WMRA_Opt(Matrix Jo, double detJo, vector<double> dx, vector<double> q);
 
 #endif;
